feat: algolia seed bull
