+++ .github/workflows/api-deploy-worker.yml
@@ -37,7 +37,6 @@ jobs:
       ECS_SERVICE: ${{ inputs.name}}-${{ inputs.env }}-service
       ECR_REPOSITORY: jfp-${{ inputs.name }}-${{ inputs.env }}
       ECR_SEED_REPOSITORY: jfp-${{ inputs.name }}-seed-${{ inputs.env }}
-      ECR_SEED_ALGOLIA_REPOSITORY: jfp-${{ inputs.name }}-alogolia-seed-${{ inputs.env }}
       IMAGE_TAG: ${{ github.sha }}
     runs-on: ubuntu-latest
     strategy:
@@ -159,19 +158,3 @@ jobs:
         if: contains(steps.affected-dbs.outputs.dbs, inputs.name)
         run: |
           aws ecs run-task --task-definition $ECR_SEED_REPOSITORY --cluster $ECS_CLUSTER --launch-type FARGATE --network-configuration "${{ secrets.AWS_NETWORK_CONFIG }}"
-
-      - name: Get affected alogolia seed
-        run: echo "dockerfile=$(git diff --name-only HEAD~1 HEAD | grep 'apps/${{ inputs.name }}/db/Dockerfile.Algolia')" >> $GITHUB_OUTPUT
-        id: affected-dockerfile-alogolia
-
-      - name: Build, tag, and push the image to Amazon ECR
-        if: env == 'prod' && (contains(steps.affected-dbs.outputs.dbs, inputs.name) || contains(steps.affected-dockerfile-alogolia.outputs.dockerfile, inputs.name))
-        id: build-image-ecs-seed-algolia
-        env:
-          ECR_REGISTRY: ${{ steps.login-ecr-ecs.outputs.registry }}
-        run: |
-          # Build a docker container and push it to ECR 
-          docker build -t $ECR_REGISTRY/$ECR_SEED_ALGOLIA_REPOSITORY:$IMAGE_TAG -t $ECR_REGISTRY/$ECR_SEED_ALGOLIA_REPOSITORY:latest . -f apps/${{ inputs.name }}/db/Dockerfile.Algolia
-          echo "Pushing image to ECR..."
-          docker push -a $ECR_REGISTRY/$ECR_SEED_ALGOLIA_REPOSITORY
-          echo "image=$ECR_REGISTRY/$ECR_SEED_ALGOLIA_REPOSITORY:$IMAGE_TAG" >> $GITHUB_OUTPUT
