--- apps/api-videos/db/syncVideosToAlgolia/index.ts
@@ -1,1 +1,0 @@
-export { syncVideosToAlgolia } from './syncVideosToAlgolia'
