+++ apps/api-videos/infrastructure/locals.tf
@@ -1,10 +1,12 @@
 locals {
   port = 4004
   environment_variables = [
-    "PG_DATABASE_URL_VIDEOS",
     "ARCLIGHT_API_KEY",
     "ARCLIGHT_V3_URL",
-    "CROWDIN_API_KEY"
+    "ALGOLIA_APP_ID",
+    "ALGOLIA_API_KEY",
+    "CROWDIN_API_KEY",
+    "PG_DATABASE_URL_VIDEOS"
   ]
   service_config = {
     name           = "api-videos"
