+++ apps/api-videos/src/app/modules/algolia/algolia.consumer.ts
@@ -1,0 +1,15 @@
+import { Processor, WorkerHost } from '@nestjs/bullmq'
+import { Job } from 'bullmq'
+
+import { AlgoliaService } from './algolia.service'
+
+@Processor('api-videos-aloglia')
+export class AlgoliaConsumer extends WorkerHost {
+  constructor(private readonly algoliaService: AlgoliaService) {
+    super()
+  }
+
+  async process(job: Job): Promise<void> {
+    await this.algoliaService.syncVideosToAlgolia()
+  }
+}
